feat: full platform update - auth, UI, database, deployment configs
- Simplified 2-step auth flow (phone → OTP)
- Updated all admin/cashier screens with RTL, dark mode, responsive fixes
- Database DAOs and schema improvements
- Added deployment scripts and Vercel/Render configs
- Added node_modules to .gitignore
- Updated shared packages (auth, pos, database, shared_ui)
- CI/CD workflows for web builds

+++ apps/admin/lib/providers/purchases_providers.dart
@@ -218,7 +218,7 @@ Future<void> receivePurchaseWithDetails(
     final product = await db.productsDao.getProductById(item.productId);
     if (product != null) {
       final newStock = product.stockQty + receivedQty;
-      await db.productsDao.updateStock(item.productId, newStock);
+      await db.productsDao.updateStock(item.productId, newStock.toInt());
 
       // تسجيل حركة المخزون
       if (storeId != null) {
@@ -226,8 +226,8 @@ Future<void> receivePurchaseWithDetails(
           id: _uuid.v4(),
           productId: item.productId,
           storeId: storeId,
-          qty: receivedQty,
-          previousQty: product.stockQty,
+          qty: receivedQty.toDouble(),
+          previousQty: product.stockQty.toDouble(),
           purchaseId: purchaseId,
         );
       }
